+++ crate/model/src/common/graphviz_dot_theme.rs
@@ -1,5 +1,13 @@
+pub use graph_style::GraphStyle;
+
+mod graph_style;
+
 #[derive(Clone, Debug, PartialEq)]
 pub struct GraphvizDotTheme {
+    /// Style of graph to render.
+    pub graph_style: GraphStyle,
+
+    /// The colour to use for graph edges.
     pub edge_color: &'static str,
 
     // Node attributes:
@@ -44,6 +52,86 @@ pub struct GraphvizDotTheme {
 }
 
 impl GraphvizDotTheme {
+    pub fn with_graph_style(mut self, graph_style: GraphStyle) -> Self {
+        self.graph_style = graph_style;
+        self
+    }
+
+    pub fn with_edge_color(mut self, edge_color: &'static str) -> Self {
+        self.edge_color = edge_color;
+        self
+    }
+
+    pub fn with_node_text_color(mut self, node_text_color: &'static str) -> Self {
+        self.node_text_color = node_text_color;
+        self
+    }
+
+    pub fn with_node_width(mut self, node_width: f64) -> Self {
+        self.node_width = node_width;
+        self
+    }
+
+    pub fn with_node_height(mut self, node_height: f64) -> Self {
+        self.node_height = node_height;
+        self
+    }
+
+    pub fn with_node_margin_x(mut self, node_margin_x: f64) -> Self {
+        self.node_margin_x = node_margin_x;
+        self
+    }
+
+    pub fn with_node_margin_y(mut self, node_margin_y: f64) -> Self {
+        self.node_margin_y = node_margin_y;
+        self
+    }
+
+    pub fn with_plain_text_color(mut self, plain_text_color: &'static str) -> Self {
+        self.plain_text_color = plain_text_color;
+        self
+    }
+
+    pub fn with_emoji_point_size(mut self, emoji_point_size: u32) -> Self {
+        self.emoji_point_size = emoji_point_size;
+        self
+    }
+
+    pub fn with_node_point_size(mut self, node_point_size: u32) -> Self {
+        self.node_point_size = node_point_size;
+        self
+    }
+
+    pub fn with_tag_width(mut self, tag_width: f64) -> Self {
+        self.tag_width = tag_width;
+        self
+    }
+
+    pub fn with_tag_height(mut self, tag_height: f64) -> Self {
+        self.tag_height = tag_height;
+        self
+    }
+
+    pub fn with_tag_margin_x(mut self, tag_margin_x: f64) -> Self {
+        self.tag_margin_x = tag_margin_x;
+        self
+    }
+
+    pub fn with_tag_margin_y(mut self, tag_margin_y: f64) -> Self {
+        self.tag_margin_y = tag_margin_y;
+        self
+    }
+
+    pub fn with_tag_point_size(mut self, tag_point_size: u32) -> Self {
+        self.tag_point_size = tag_point_size;
+        self
+    }
+
+    pub fn with_tag_classes(mut self, tag_classes: &'static str) -> Self {
+        self.tag_classes = tag_classes;
+        self
+    }
+
     pub fn edge_color(&self) -> &str {
         self.edge_color
     }
@@ -108,6 +196,7 @@ impl GraphvizDotTheme {
 impl Default for GraphvizDotTheme {
     fn default() -> Self {
         Self {
+            graph_style: GraphStyle::default(),
             edge_color: "#333333",
             node_text_color: "#111111",
             node_width: 0.3,