+++ crate/model/src/info_graph/node_info.rs
@@ -14,6 +14,27 @@ pub struct NodeInfo {
 }
 
 impl NodeInfo {
+    /// Returns a new `NodeInfo`.
+    pub fn new(name: String) -> Self {
+        Self {
+            emoji: None,
+            name,
+            desc: None,
+        }
+    }
+
+    /// Sets the emoji to display.
+    pub fn with_emoji(mut self, emoji: Option<String>) -> Self {
+        self.emoji = emoji;
+        self
+    }
+
+    /// Sets the plain text description.
+    pub fn with_desc(mut self, desc: Option<String>) -> Self {
+        self.desc = desc;
+        self
+    }
+
     /// Returns the emoji to display.
     pub fn emoji(&self) -> Option<&str> {
         self.emoji.as_deref()