+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## 0.1.1 (2024-02-17)
+## unreleased
 
 * Split `dot_ix` library and `dot_ix_playground` binary.
+* Add `InfoGraph::builder` and `InfoGraphBuilder`.
+* Add `NodeId::new`.
 
 
 ## 0.1.0 (2024-02-04)